introduce flux subpackage
